requirements.txt
streamlit
numpy
pandas
matplotlib
scikit-learn
pymongo
bcrypt
python-dotenv